feat: add selecting priority in add-todo-form
