+++ components/add-todo-form/types/form-model.ts
@@ -1,0 +1,6 @@
+import { TodoPriority } from "@/shared/constants";
+
+export interface AddTodoFormModel {
+  title: string;
+  priority: TodoPriority;
+}
